added tests for the new setTypes approach.

+++ qa/Test_dm.py
@@ -29,9 +29,9 @@ import dm
 ##############
 class Test_dm( unittest.TestCase ) :
 
-  logging.basicConfig( format='%(levelname)s:%(message)s', level=logging.DEBUG )
+  #logging.basicConfig( format='%(levelname)s:%(message)s', level=logging.DEBUG )
   #logging.basicConfig( format='%(levelname)s:%(message)s', level=logging.INFO )
-  #logging.basicConfig( format='%(levelname)s:%(message)s', level=logging.WARNING )
+  logging.basicConfig( format='%(levelname)s:%(message)s', level=logging.WARNING )
 
   PRINT_STOP    = False
   COMPARE_PROGS = True
@@ -196,7 +196,8 @@ class Test_dm( unittest.TestCase ) :
     #clean up testing
 
     IRDB.close()
-    os.remove( testDB )
+    if os.path.exists( testDB ) :
+      os.remove( testDB )
 
 
   ##############
@@ -2483,6 +2484,10 @@ class Test_dm( unittest.TestCase ) :
 #    argDict[ 'EFF' ]                      = 2
 #
 #    return argDict
+
+
+if __name__ == "__main__" :
+  unittest.main()
 
 
 #########
